Bug fix
When a text answer response has been run by the program and the user imports data before answering the question, the answer will be cleared. This issue still persists. However, to access the save and load buttons the user used to have to run the program. Sometimes the game will start with a text response question and if the user chose to import data then the correct answer would not be collected from the imported data. 

Now on the first button press, the save/load buttons appear but the program does not run the question. This allows for the importing of data to be the user's first action without the answer being lost.